requirements.txt
importlib_metadata==4.6.4
jira==3.6.0
pytest==8.0.0
PyYAML==6.0.1
questionary==2.0.1
rich==13.7.0
setuptools==59.6.0
Sphinx==7.2.6
